requirements.txt
contourpy==1.3.3
cycler==0.12.1
fonttools==4.61.1
kiwisolver==1.4.9
matplotlib==3.10.8
narwhals==2.16.0
numpy==2.4.2
packaging==26.0
pillow==12.1.0
pillow_heif==1.2.0
plotly==6.5.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
PyYAML==6.0.3
setuptools==80.10.2
six==1.17.0
tqdm==4.67.3
wheel==0.46.3
